feat: interface service layer
